+++ src/internal/tester/resource_owners.go
@@ -72,6 +72,18 @@ func SecondaryM365UserID(t *testing.T) string {
 	return strings.ToLower(cfg[TestCfgSecondaryUserID])
 }
 
+// TertiaryM365UserID returns an userID string representing the m365UserID
+// described by either the env var CORSO_TERTIARY_M365_TEST_USER_ID, the
+// corso_test.toml config file or the default value (in that order of priority).
+// The default is a last-attempt fallback that will only work on alcion's
+// testing org.
+func TertiaryM365UserID(t *testing.T) string {
+	cfg, err := readTestConfig()
+	require.NoError(t, err, "retrieving tertiary m365 user id from test configuration", clues.ToCore(err))
+
+	return strings.ToLower(cfg[TestCfgTertiaryUserID])
+}
+
 // LoadTestM365SiteID returns a siteID string representing the m365SiteID
 // described by either the env var CORSO_M365_LOAD_TEST_SITE_ID, the
 // corso_test.toml config file or the default value (in that order of priority).