+++ src/pkg/selectors/onedrive_test.go
@@ -74,9 +74,9 @@ func (suite *OneDriveSelectorSuite) TestOneDriveSelector_AllData() {
 				scopeMustHave(
 					t,
 					OneDriveScope(scope),
-					map[categorizer]string{
-						OneDriveItem:   AnyTgt,
-						OneDriveFolder: AnyTgt,
+					map[categorizer][]string{
+						OneDriveItem:   Any(),
+						OneDriveFolder: Any(),
 					},
 				)
 			}
@@ -106,9 +106,9 @@ func (suite *OneDriveSelectorSuite) TestOneDriveSelector_Include_AllData() {
 		scopeMustHave(
 			t,
 			OneDriveScope(sc),
-			map[categorizer]string{
-				OneDriveItem:   AnyTgt,
-				OneDriveFolder: AnyTgt,
+			map[categorizer][]string{
+				OneDriveItem:   Any(),
+				OneDriveFolder: Any(),
 			},
 		)
 	}
@@ -136,9 +136,9 @@ func (suite *OneDriveSelectorSuite) TestOneDriveSelector_Exclude_AllData() {
 		scopeMustHave(
 			t,
 			OneDriveScope(sc),
-			map[categorizer]string{
-				OneDriveItem:   AnyTgt,
-				OneDriveFolder: AnyTgt,
+			map[categorizer][]string{
+				OneDriveItem:   Any(),
+				OneDriveFolder: Any(),
 			},
 		)
 	}
@@ -270,7 +270,7 @@ func (suite *OneDriveSelectorSuite) TestOneDriveCategory_PathValues() {
 
 	expected := map[categorizer][]string{
 		OneDriveFolder: {"dir1/dir2"},
-		OneDriveItem:   {fileName, shortRef},
+		OneDriveItem:   {fileName, shortRef, fileName + "-id"},
 	}
 
 	ent := details.DetailsEntry{