+++ CHANGELOG.md
@@ -25,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed failure when downloading huge amount of attachments
 - Graph API requests that return an ECONNRESET error are now retried.
 
+### Known Issues
+- Restoring a OneDrive or SharePoint file with the same name as a file with that name as its M365 ID may restore both items.
+
 ## [v0.6.1] (beta) - 2023-03-21
 
 ### Added